Add documentation and move write_to_dag_file to helpers
